+++ host/include/matrix.h
@@ -4,14 +4,14 @@
 typedef struct matrix {
     int rows;
     int cols;
-    double** vals;
+    float** vals;
 } matrix_t;
 
 matrix_t* create_matrix(int rows, int cols);
 
 matrix_t* create_matrix_identity(int dim);
 
-matrix_t* create_matrix_random(int rows, int cols, double start, double end);
+matrix_t* create_matrix_random(int rows, int cols, float start, float end);
 
 void destroy_matrix(matrix_t* m);
 
@@ -35,7 +35,7 @@ matrix_t* row_sum_matrix(matrix_t* m);
 
 matrix_t* transpose_matrix(matrix_t* m);
 
-void apply_matrix(double (*apply)(double), matrix_t* m);
+void apply_matrix(float (*apply)(float), matrix_t* m);
 
 matrix_t* copy_submatrix(matrix_t* m, int start_row, int end_row, int start_col, int end_col);
